requirements.txt
numpy
pandas
pillow
matplotlib
scikit-image
scikit-learn
scipy
torch
torchvision
open_clip_torch
lpips
